+++ regression/reference/char-cxx/wrapfchar.f
@@ -319,23 +319,23 @@ module char_mod
     ! end get_const_char_ptr_raw
 
     ! ----------------------------------------
-    ! Function:  const char *getConstCharPtrAsCopyArg
+    ! Function:  const char *getConstCharPtrAsCopyArg +deref(copy)+funcarg
     ! Statement: c_function_char*
     ! start c_get_const_char_ptr_as_copy_arg
     interface
         function c_get_const_char_ptr_as_copy_arg() &
-                result(SHT_rv) &
+                result(output) &
                 bind(C, name="CHA_getConstCharPtrAsCopyArg")
             use iso_c_binding, only : C_PTR
             implicit none
-            type(C_PTR) :: SHT_rv
+            type(C_PTR) :: output
         end function c_get_const_char_ptr_as_copy_arg
     end interface
     ! end c_get_const_char_ptr_as_copy_arg
 
     ! ----------------------------------------
-    ! Function:  const char *getConstCharPtrAsCopyArg
-    ! Statement: f_function_char*_buf_arg
+    ! Function:  const char *getConstCharPtrAsCopyArg +deref(copy)+funcarg
+    ! Statement: f_function_char*_buf_funcarg_copy
     ! start c_get_const_char_ptr_as_copy_arg_bufferify
     interface
         subroutine c_get_const_char_ptr_as_copy_arg_bufferify(output, &
@@ -1029,8 +1029,8 @@ contains
 #endif
 
     ! ----------------------------------------
-    ! Function:  const char *getConstCharPtrAsCopyArg
-    ! Statement: f_function_char*_buf_arg
+    ! Function:  const char *getConstCharPtrAsCopyArg +deref(copy)+funcarg
+    ! Statement: f_function_char*_buf_funcarg_copy
     !>
     !! \brief return a 'const char *' as argument
     !!