+++ regression/reference/strings-cfi/wrapfstrings.f
@@ -78,8 +78,8 @@ module strings_mod
     end interface
 
     ! ----------------------------------------
-    ! Function:  const string getConstStringAsArg
-    ! Statement: f_function_string_cfi_arg
+    ! Function:  const string getConstStringAsArg +deref(copy)+funcarg
+    ! Statement: f_function_string_cfi_funcarg_copy
     interface
         subroutine c_get_const_string_as_arg_CFI(output) &
                 bind(C, name="STR_getConstStringAsArg_CFI")
@@ -152,21 +152,21 @@ module strings_mod
     end interface
 
     ! ----------------------------------------
-    ! Function:  const string &getConstStringRefAsArg
+    ! Function:  const string &getConstStringRefAsArg +deref(copy)+funcarg
     ! Statement: c_function_string&
     interface
         function c_get_const_string_ref_as_arg() &
-                result(SHT_rv) &
+                result(output) &
                 bind(C, name="STR_getConstStringRefAsArg")
             use iso_c_binding, only : C_PTR
             implicit none
-            type(C_PTR) :: SHT_rv
+            type(C_PTR) :: output
         end function c_get_const_string_ref_as_arg
     end interface
 
     ! ----------------------------------------
-    ! Function:  const string &getConstStringRefAsArg
-    ! Statement: f_function_string&_cfi_arg
+    ! Function:  const string &getConstStringRefAsArg +deref(copy)+funcarg
+    ! Statement: f_function_string&_cfi_funcarg_copy
     interface
         subroutine c_get_const_string_ref_as_arg_CFI(output) &
                 bind(C, name="STR_getConstStringRefAsArg_CFI")
@@ -856,8 +856,8 @@ contains
     end function get_const_string_len
 
     ! ----------------------------------------
-    ! Function:  const string getConstStringAsArg
-    ! Statement: f_function_string_cfi_arg
+    ! Function:  const string getConstStringAsArg +deref(copy)+funcarg
+    ! Statement: f_function_string_cfi_funcarg_copy
     !>
     !! \brief return a 'const string' as argument
     !!
@@ -916,8 +916,8 @@ contains
     end function get_const_string_ref_len
 
     ! ----------------------------------------
-    ! Function:  const string &getConstStringRefAsArg
-    ! Statement: f_function_string&_cfi_arg
+    ! Function:  const string &getConstStringRefAsArg +deref(copy)+funcarg
+    ! Statement: f_function_string&_cfi_funcarg_copy
     !>
     !! \brief return a 'const string&' as argument
     !!